Added controller actions to match service methods.
Added controller actions to the Authentication and Users controllers to match that of the corresponding services. Need to add a couple more for some 2FA stuff though.
Added more request DTOs. Need to consolidate those back into some base class since most of them are pretty much the same.
Added an extension to the Controller to get the current user.
Slight name changes to some methods.