UserService and API adjustments.
Consolidated Requests and Responses into a base class.
Adjusted the Controller Actions to pretty much always return a form of BaseResponse object. This gives us some more granularity and flexibility.
Created an override for the UserService.CreateUser method, so we can create a user without having to send a confirmation email. Saves the need of the extra parameters required by the email sender.
Moved the Email Confirmation actions above the Password Reset actions. It just seemed like the right thing to do...
Added a method to determine whether or not the currently logged in user is the same as the one detailed in the Request. Has the ability for administrator override.