Added Map API, persistence layer and Hangfire
Added a REST API for the Status data.
Moved the DbContext into a separate Persistence project.
Added Hangfire for Task Scheduling the VATSIM Status data updates.
Stripped out the VatsimStatusDataParser and improved the performance.
Added extra logging to the StatusService.
Accounted for eager loading in the StatusService.
Added attribute to define a particular type as being specific to a certain Online Flight Simulation Network.
Changed string based IDs to integers.
Added a JSON converter for the SquawkCode value type.