Added dummy Auth project
Added a dummy ASP.NET Core application for auth. Was supposed to be dedicated IdentityServer, but it'll be best if it's combined with the Identity API.
Just keeping it in because i started working on it already, and i need a reference to work off while i move everything over.